style(sources): force -Werror=useless-cast and old-style-cast
and fix all warnings for #1950

+++ src/reports/forecast.cpp
@@ -71,7 +71,7 @@ wxString mmReportForecast::getHTMLText()
     mm_html_template report(this->m_template);
     report(L"REPORTNAME") = this->getReportTitle();
     report(L"CONTENTS") = contents;
-    report(L"GRAND") = wxString::Format("%ld", (long)amount_by_day.size());
+    report(L"GRAND") = wxString::Format("%zu", amount_by_day.size());
     report(L"HTMLSCALE") = wxString::Format("%d", Option::instance().getHtmlFontSize());
 
     wxString out = wxEmptyString;
